Component generation in seperate factory
